Fix to_visual algorithm broken after first call
This change fixes an issue where to_visual was only returning the
correctly formatted bidi result the first time it was called after a
server restart (or anything else that would re-initialize Bidi’s global
$weakHashMap variable).

The root cause of this issue was an unless block with improper syntax.
Ruby doesn’t allow braces to be used to define multi-line unless
blocks. Instead, the ‘end’ keyword must be used. This caused the ruby
interpreter to think that the end of the WeakHashMap’s [] method
definition was the real end of the offending unless block. Due to this,
on any WeakHashMap ‘hit’ the code that actually assigns ret_value was
always skipped and the delete method definition was returned instead.
This caused every letter to have bidiType nil in bidi.rb’s
to_paragraphs method which messed up the bidi algorithm.

This is also the reason why an extra end keyword was previously needed
at the bottom of the file. Were it not there, a SyntaxError would have
been raised.

I manually tested this change by testing that to_visual returned the
correctly formatted bidi string on its first, second, and subsequent
calls with the input string "משפט עם עברית ו-English. מספרים: 12345
(וגם כל מיני סימני פיסוק) וגם סימן קריאה!” taken from this thread
https://github.com/prawnpdf/prawn/issues/514